Created a selectCell method in the PaintedGrid (#282)
